Extract matches function
