+++ lib/share/shared.js
@@ -2,9 +2,10 @@ const EventEmitter = require('events')
 const fs = require('fs')
 const separator = require('path').sep
 const debug = require('debug')('slsk:shared:i')
+const matches = require('./matches.js')
 
 module.exports = class Shared extends EventEmitter {
-  constructor (folders) {
+  constructor () {
     super()
     this.files = []
 
@@ -43,13 +44,6 @@ module.exports = class Shared extends EventEmitter {
   }
 
   search (query) {
-    return this.files
-      .filter(it => query
-        .split(' ')
-        .map(criterion => criterion.startsWith('-')
-          ? it.key.indexOf(criterion.substring(1)) === -1
-          : it.key.indexOf(criterion) !== -1
-        ).reduce((a, b) => a && b, true)
-      )
+    return this.files.filter(it => matches(it.key, query))
   }
 }
