+++ lib/slsk-client.js
@@ -10,7 +10,7 @@ const DefaultPeer = require('./peer/default-peer.js')
 const DistributedPeer = require('./peer/distributed-peer.js')
 const downloadPeerFile = require('./peer/download-peer-file.js')
 const Listen = require('./listen.js')
-const Shared = require('./shared.js')
+const Shared = require('./share/shared.js')
 let stack = require('./stack')
 
 let server
